refactor(mobile): split local import cycles
