+++ apps/mobile/src/screens/Address/components/MultiAssets/ProtocolList.tsx
@@ -15,7 +15,7 @@ import { GestureDetector, RefreshControl } from 'react-native-gesture-handler';
 import { getItemId } from '@/screens/Home/utils/listRenderId';
 import { KeyringAccountWithAlias } from '@/hooks/account';
 import useLoadMoreData from './hooks/useLoadMoreData';
-import { TAB_HEADER_FULL_HEIGHT, TabName } from './TabsMultiAssets';
+import { HomeTabName as TabName } from '@/hooks/navigation';
 import {
   ListRenderFooter as ListRenderFooterComponent,
   ListRenderSeparator,
