+++ apps/mobile/src/screens/Transaction/components/Actions/ApproveNFTCollection.tsx
@@ -23,7 +23,8 @@ import { ApproveNFTRequireData } from '@rabby-wallet/rabby-action';
 import { useMemoizedFn } from 'ahooks';
 import { unionBy } from 'lodash';
 import { useTranslation } from 'react-i18next';
-import { AddressItemInDetail, TxStatusItem } from '../../HistoryDetailScreen';
+import { AddressItemInDetail } from '../../HistoryDetailScreen';
+import { TxStatusItem } from '../TxStatusItem';
 import { HistoryItemIcon } from '../HistoryItemIcon';
 import { HistoryItemCateType } from '../type';
 import { RevokeNFTCollectionBtn } from './components/RevokeNFTCollectionBtn';
