+++ apps/mobile/src/screens/Transaction/components/Actions/DeployContract.tsx
@@ -17,7 +17,8 @@ import { openTxExternalUrl } from '@/utils/transaction';
 import { useMemoizedFn } from 'ahooks';
 import { unionBy } from 'lodash';
 import { useTranslation } from 'react-i18next';
-import { AddressItemInDetail, TxStatusItem } from '../../HistoryDetailScreen';
+import { AddressItemInDetail } from '../../HistoryDetailScreen';
+import { TxStatusItem } from '../TxStatusItem';
 import { ellipsisAddress } from '@/utils/address';
 import { formatIntlTimestamp } from '@/utils/time';
 import { formatAmount } from '@/utils/number';
