+++ apps/mobile/src/setup-app-before-render.runtime.ts
@@ -55,10 +55,11 @@ import {
 } from './store/curve24h';
 import useProtocolListStore from './store/protocols';
 import { useAppChainStore } from './store/appchain';
-import addressBalanceStore, {
+import addressBalanceStore from './store/balance';
+import {
   ensureAccountBalanceSelectionLifecycle,
   startProcessAccountBalanceEvents,
-} from './store/balance';
+} from './store/balanceAccountSelection';
 import * as apisAutoLock from './core/apis/autoLock';
 import { startWatchLayoutChange } from './hooks/useAppLayout';
 import { startCareAppNotificationPermissions } from './hooks/appNotification';
